♻️(malware) extract shared backend record-management into BaseBackend
The analyse_file, reschedule_processing_task, launch_next_analysis,
_update_detection_file_hash, delete_detection, failed_analysis and
succeed_analysis methods were duplicated verbatim between the ICAP and
JCOP backends, differing only in which Celery task's .delay they
invoked. Hoist them into BaseBackend and parametrize the task via a
_analysis_task class attribute that each backend sets to its
backend-specific task (scan_file_async / analyse_file_async).

Behavior-neutral: existing tests patch .delay on the task object
itself, so the new class-attribute reference keeps mocking working.
