+++ src/lasuite/malware_detection/backends/base.py
@@ -1,14 +1,25 @@
 """malware_detection backend base module to implement."""
 
-from abc import ABC, abstractmethod
+import logging
+from abc import ABC
 
+from django.core.files.storage import default_storage
+from django.db import transaction
+from django.db.models import Q
 from django.utils.module_loading import import_string
 
-from ..models import MalwareDetection
+from ..enums import ReportStatus
+from ..models import MalwareDetection, MalwareDetectionStatus
+
+logger = logging.getLogger(__name__)
 
 
 class BaseBackend(ABC):
     """Base class for all malware detection backends."""
+
+    #: The Celery task class responsible for analysing a file. Subclasses
+    #: assign this to their backend-specific task (e.g. ``scan_file_async``).
+    _analysis_task: object
 
     def __init__(self, *, callback_path: str, max_processing_files: int = 15, **kwargs):
         """Initialize the backend."""
@@ -29,14 +40,124 @@ class BaseBackend(ABC):
         """Get the callback function."""
         return import_string(self.callback_path)
 
-    @abstractmethod
     def analyse_file(self, file_path: str, **kwargs) -> None:
-        """Analyse a file and call the callback with the result."""
+        """Trigger a process to analyse a file using the backend service."""
+        if not default_storage.exists(file_path):
+            raise FileNotFoundError(f"File {file_path} not found")
 
-    @abstractmethod
-    def launch_next_analyses(self) -> None:
-        """Launch the pending analyses."""
+        if self.max_processing_files == 0:
+            # if max_processing_files is 0, we don't want to create a detection record
+            # we just want to analyse the file asynchronously
+            self._analysis_task.delay(
+                file_path,
+                **kwargs,
+            )
+            return
 
-    @abstractmethod
+        if not MalwareDetection.objects.filter(
+            path=file_path,
+            status__in=[MalwareDetectionStatus.PENDING, MalwareDetectionStatus.PROCESSING],
+        ).exists():
+            MalwareDetection.objects.create(
+                path=file_path,
+                status=MalwareDetectionStatus.PENDING,
+                backend=self.backend_name,
+                parameters=kwargs,
+            )
+
+        self.launch_next_analyses()
+
     def reschedule_processing_task(self, malware_detection_record: MalwareDetection) -> None:
         """Reschedule the processing task for a malware detection record."""
+        if malware_detection_record.status != MalwareDetectionStatus.PROCESSING:
+            return
+
+        if malware_detection_record.backend != self.backend_name:
+            return
+
+        # assert first the file still exists in the system
+        if not default_storage.exists(malware_detection_record.path):
+            logger.info("File %s not found when rescheduling processing task", malware_detection_record.path)
+            malware_detection_record.delete()
+            return
+
+        self._analysis_task.delay(
+            file_path=malware_detection_record.path,
+            file_hash=malware_detection_record.file_hash,
+            **malware_detection_record.parameters,
+        )
+
+    def launch_next_analyses(self) -> None:
+        """Launch the pending analyses until all processing slots are filled."""
+        free_slots = (
+            self.max_processing_files
+            - MalwareDetection.objects.filter(status=MalwareDetectionStatus.PROCESSING).count()
+        )
+        if free_slots <= 0:
+            return
+
+        with transaction.atomic():
+            pending_detections = list(
+                MalwareDetection.objects.select_for_update()
+                .filter(status=MalwareDetectionStatus.PENDING)
+                .order_by("created_at")[:free_slots]
+            )
+            if not pending_detections:
+                return
+            MalwareDetection.objects.filter(pk__in=[detection.pk for detection in pending_detections]).update(
+                status=MalwareDetectionStatus.PROCESSING
+            )
+
+        for detection in pending_detections:
+            self._analysis_task.delay(
+                detection.path,
+                **detection.parameters,
+            )
+
+    def _update_detection_file_hash(self, file_path: str, file_hash: str) -> None:
+        """Update the file hash for a detection record."""
+        MalwareDetection.objects.filter(
+            path=file_path,
+        ).filter(~Q(file_hash=file_hash)).update(file_hash=file_hash)
+
+    def delete_detection(self, file_path: str) -> None:
+        """Delete the detection records for a path."""
+        deleted, _ = MalwareDetection.objects.filter(path=file_path).delete()
+        if not deleted:
+            logger.warning("Detection %s not found", file_path)
+
+    def failed_analysis(
+        self,
+        file_path: str,
+        file_hash: str | None,
+        error_code: int,
+        error_msg: str,
+        status: ReportStatus | None = None,
+        **kwargs,
+    ) -> None:
+        """Handle a failed analysis."""
+        updated = MalwareDetection.objects.filter(path=file_path).update(
+            status=MalwareDetectionStatus.FAILED,
+            error_code=error_code,
+            error_msg=error_msg,
+        )
+        if not updated:
+            logger.warning("Detection %s not found", file_path)
+
+        self.launch_next_analyses()
+        self.callback(
+            file_path,
+            status if status is not None else ReportStatus.UNKNOWN,
+            error_info={
+                "error": error_msg,
+                "error_code": error_code,
+                "file_hash": file_hash,
+            },
+            **kwargs,
+        )
+
+    def succeed_analysis(self, file_path: str, **kwargs) -> None:
+        """Handle a successful analysis."""
+        self.delete_detection(file_path)
+        self.launch_next_analyses()
+        self.callback(file_path, ReportStatus.SAFE, error_info={}, **kwargs)
