+++ src/lasuite/malware_detection/backends/icap.py
@@ -1,28 +1,24 @@
 """Module contains the ICAP backend for the malware detection system."""
 
 import hashlib
-import logging
 import ssl
 from http import HTTPStatus
 from pathlib import Path
 from urllib.parse import quote
 
 from django.core.files.storage import default_storage
-from django.db import transaction
-from django.db.models import Q
 from icap import IcapClient
 from icap.exception import IcapServerError
 
 from ..enums import ReportStatus
-from ..models import MalwareDetection, MalwareDetectionStatus
 from ..tasks.icap import scan_file_async
 from .base import BaseBackend
-
-logger = logging.getLogger(__name__)
 
 
 class ICAPBackend(BaseBackend):
     """A backend that uses an ICAP server to detect malware."""
+
+    _analysis_task = scan_file_async
 
     def __init__(  # noqa: PLR0913, PLR0917
         self,
@@ -88,74 +84,6 @@ class ICAPBackend(BaseBackend):
         """
         resource = "/" + quote(filename, safe="") if filename else "/scan"
         return f"POST {resource} HTTP/1.1\r\nHost: file-scan\r\n\r\n".encode()
-
-    def analyse_file(self, file_path: str, **kwargs) -> None:
-        """Trigger a process to analyse a file using the ICAP service."""
-        if not default_storage.exists(file_path):
-            raise FileNotFoundError(f"File {file_path} not found")
-
-        if self.max_processing_files == 0:
-            # if max_processing_files is 0, we don't want to create a detection record
-            # we just want to analyse the file asynchronously
-            scan_file_async.delay(
-                file_path,
-                **kwargs,
-            )
-            return
-
-        MalwareDetection.objects.create(
-            path=file_path,
-            status=MalwareDetectionStatus.PENDING,
-            backend=self.backend_name,
-            parameters=kwargs,
-        )
-
-        self.launch_next_analysis()
-
-    def reschedule_processing_task(self, malware_detection_record: MalwareDetection) -> None:
-        """Reschedule the processing task for a malware detection record."""
-        if malware_detection_record.status != MalwareDetectionStatus.PROCESSING:
-            return
-
-        if malware_detection_record.backend != self.backend_name:
-            return
-
-        # assert first the file still exists in the system
-        if not default_storage.exists(malware_detection_record.path):
-            logger.info("File %s not found when rescheduling processing task", malware_detection_record.path)
-            malware_detection_record.delete()
-            return
-
-        scan_file_async.delay(
-            file_path=malware_detection_record.path,
-            file_hash=malware_detection_record.file_hash,
-            **malware_detection_record.parameters,
-        )
-
-    def launch_next_analysis(self) -> None:
-        """Launch the next pending analysis."""
-        if (
-            MalwareDetection.objects.filter(status=MalwareDetectionStatus.PROCESSING).count()
-            >= self.max_processing_files
-        ):
-            return
-
-        with transaction.atomic():
-            first_pending_detection = (
-                MalwareDetection.objects.select_for_update()
-                .filter(status=MalwareDetectionStatus.PENDING)
-                .order_by("created_at")
-                .first()
-            )
-            if first_pending_detection is None:
-                return
-            first_pending_detection.status = MalwareDetectionStatus.PROCESSING
-            first_pending_detection.save(update_fields=["status"])
-
-        scan_file_async.delay(
-            first_pending_detection.path,
-            **first_pending_detection.parameters,
-        )
 
     def scan_file(self, file_path: str, file_hash: str | None = None, **kwargs) -> None:
         """Scan a file with the ICAP server and report the result."""
@@ -225,57 +153,3 @@ class ICAPBackend(BaseBackend):
             if separator and key.lower() == "threat":
                 return value
         return infection
-
-    def _update_detection_file_hash(self, file_path: str, file_hash: str) -> None:
-        """Update the file hash for a detection record."""
-        MalwareDetection.objects.filter(
-            path=file_path,
-        ).filter(~Q(file_hash=file_hash)).update(file_hash=file_hash)
-
-    def delete_detection(self, file_path: str) -> None:
-        """Delete a detection record."""
-        try:
-            detection = MalwareDetection.objects.get(path=file_path)
-        except MalwareDetection.DoesNotExist:
-            logger.warning("Detection %s not found", file_path)
-            return
-        else:
-            detection.delete()
-
-    def failed_analysis(
-        self,
-        file_path: str,
-        file_hash: str,
-        error_code: int,
-        error_msg: str,
-        status: ReportStatus | None = None,
-        **kwargs,
-    ) -> None:
-        """Handle a failed analysis."""
-        try:
-            detection = MalwareDetection.objects.get(path=file_path)
-        except MalwareDetection.DoesNotExist:
-            logger.warning("Detection %s not found", file_path)
-        else:
-            detection.status = MalwareDetectionStatus.FAILED
-            detection.error_code = error_code
-            detection.error_msg = error_msg
-            detection.save(update_fields=["status", "error_code", "error_msg"])
-
-        self.launch_next_analysis()
-        self.callback(
-            file_path,
-            status if status is not None else ReportStatus.UNKNOWN,
-            error_info={
-                "error": error_msg,
-                "error_code": error_code,
-                "file_hash": file_hash,
-            },
-            **kwargs,
-        )
-
-    def succeed_analysis(self, file_path: str, **kwargs) -> None:
-        """Handle a successful analysis."""
-        self.delete_detection(file_path)
-        self.launch_next_analysis()
-        self.callback(file_path, ReportStatus.SAFE, error_info={}, **kwargs)
