+++ tests/malware_detection/backends/test_icap_backend.py
@@ -37,12 +37,18 @@ def icap_generate_file_path():
 
 
 @pytest.fixture
-def icap_backend():
+def icap_backend_parameters():
+    """Return the parameters for the ICAP backend."""
+    return {
+        "server_address": "icap.example.com",
+        "callback_path": "tests.malware_detection.backends.test_icap_backend.icap_callback",
+    }
+
+
+@pytest.fixture
+def icap_backend(icap_backend_parameters):
     """Create an ICAP backend instance with test configuration."""
-    return icap.ICAPBackend(
-        server_address="icap.example.com",
-        callback_path="tests.malware_detection.backends.test_icap_backend.icap_callback",
-    )
+    return icap.ICAPBackend(**icap_backend_parameters)
 
 
 @pytest.fixture
@@ -436,29 +442,75 @@ def test_icap_backend_failed_analysis_without_record(icap_backend):
     )
 
 
-def test_icap_backend_launch_next_analysis(icap_generate_file_path, icap_backend):
-    """Launch a new analysis when calling launch_next_analysis."""
+def test_icap_backend_succeed_analysis_with_duplicate_paths(icap_backend):
+    """A successful analysis should complete even when the path has duplicate records."""
+    file_path = "file.txt"
+    factories.MalwareDetectionFactory.create_batch(2, path=file_path, status=MalwareDetectionStatus.PROCESSING)
+
+    icap_backend.succeed_analysis(file_path)
+
+    assert not MalwareDetection.objects.filter(path=file_path).exists()
+    icap_callback.assert_called_once_with(file_path, ReportStatus.SAFE, error_info={})
+
+
+def test_icap_backend_failed_analysis_with_duplicate_paths(icap_backend):
+    """A failed analysis should complete even when the path has duplicate records."""
+    file_path = "file.txt"
+    factories.MalwareDetectionFactory.create_batch(2, path=file_path, status=MalwareDetectionStatus.PROCESSING)
+
+    icap_backend.failed_analysis(file_path, "hash123", 5000, "malware detected", ReportStatus.UNSAFE)
+
+    assert (
+        MalwareDetection.objects.filter(
+            path=file_path, status=MalwareDetectionStatus.FAILED, error_code=5000, error_msg="malware detected"
+        ).count()
+        == 2
+    )
+    icap_callback.assert_called_once_with(
+        file_path,
+        ReportStatus.UNSAFE,
+        error_info={"error": "malware detected", "error_code": 5000, "file_hash": "hash123"},
+    )
+
+
+def test_icap_backend_launch_next_analyses(icap_generate_file_path, icap_backend):
+    """Launch all the pending analyses when calling launch_next_analyses."""
     file_path, _ = icap_generate_file_path
     factories.MalwareDetectionFactory(status=MalwareDetectionStatus.PENDING, path=file_path, parameters={"foo": "bar"})
 
+    # Add several more pending detections
     factories.MalwareDetectionFactory.create_batch(10, status=MalwareDetectionStatus.PENDING)
 
     with (
         mock.patch.object(scan_file_async, "delay") as scan_file_async_mock,
     ):
-        icap_backend.launch_next_analysis()
-        scan_file_async_mock.assert_called_once_with(
+        icap_backend.launch_next_analyses()
+        assert scan_file_async_mock.call_count == 11
+        assert scan_file_async_mock.call_args_list[0] == mock.call(
             file_path,
             foo="bar",
         )
 
 
-def test_icap_backend_launch_next_analysis_no_pending_detection(icap_backend):
+def test_icap_backend_launch_next_analyses_fills_all_free_slots(icap_backend_parameters):
+    """launch_next_analyses should dispatch pending analyses until all slots are filled."""
+    backend = icap.ICAPBackend(**icap_backend_parameters, max_processing_files=3)
+    factories.MalwareDetectionFactory.create_batch(5, status=MalwareDetectionStatus.PENDING)
+
+    with mock.patch.object(scan_file_async, "delay") as scan_file_async_mock:
+        backend.launch_next_analyses()
+
+    assert scan_file_async_mock.call_count == 3
+    assert MalwareDetection.objects.filter(status=MalwareDetectionStatus.PROCESSING).count() == 3
+    assert MalwareDetection.objects.filter(status=MalwareDetectionStatus.PENDING).count() == 2
+
+
+def test_icap_backend_launch_next_analyses_no_pending_detection(icap_backend):
     """When there is no pending detection, the method does nothing."""
     with (
         mock.patch.object(scan_file_async, "delay") as scan_file_async_mock,
     ):
-        icap_backend.launch_next_analysis()
+        icap_backend.launch_next_analyses()
         scan_file_async_mock.assert_not_called()
